Updated the events-factory.js so that the id of the space is added to the event.
In the vuecal.scss there is some hard code color to update based on the space id in the class. All of this should be made that it is calculated dynamicly.

To get the colors from click up i added a function to build a color map to clickup-service.js, and a fucntion in background.js to comunicate that to the front. This last one might be deleted.