tasks are events are colord. Colormap gotten from clickup service is turned into a css style taext that is then added to the head.
On Dark colors the text becomes hard to read, so that is something that needs to be fixed. Also the color is not added right after creation. Still have to update the event factory so that it is colored. got to find a way to get the space id from newly created entries.